minitterm: add log filter
New filter to log session to a file. The file is opened in binary mode
so it can be applied before control codes removal.

File is saved to "session.log", waiting for a way to choose file name
and path.

File is opened upon filter instanciation and closed on object
destruction: this is unsafe and could result in exceptions. But it's
better than opening the file on each rx or echo.