miniterm: properly handle line timestamping
As the filter may receive an incomplete text, it cannot replace the EOL
if it is a two chars EOL (i.e. CRLF). So buffer the current text and
return an empty string. The next chunk will contain the last EOL char so
unbuffer the whole thing and do the replacement.

In addition, with an EOL different from CR, there might be some lone CR
chars to perform real carriage returns (e.g. to replace/rewrite the
current line). So rewrite the current timestamp.

Finally, if the filter receives a text ending by CR and the EOL is CRLF
and there is no extra chars after this, the line won't change until a
new char is received.